Create structure
This application allows a single player to take a quiz by reading a txt file. The quiz tracks the questions the player answered correctly. The program reads a file to populate the questions. The program utilizes 2 structures, the reading of a file, a function to open an output file and save the player data to the file.